+++ test/on2013_05/on2013_05_05_uva/UVA10594/UVA10594.java
@@ -26,7 +26,7 @@ public class UVA10594 {
         }
         int D = in.nextInt(), K = in.nextInt();
         Graph graph = new BidirectionalGraph(N);
-        for (int i = 0; i < M; i++) graph.addFlow(x[i], y[i], K, cost[i]);
+        for (int i = 0; i < M; i++) graph.addFlow(x[i], y[i], (double) K, (double) cost[i]);
         long result = MinimumCostFlow.minimumCostFlow(graph, 0, N - 1, D);
         if (result == Long.MAX_VALUE) out.printFormat("Impossible.\n");
         else out.printFormat("%d\n", result);
